Re organize tests
