+++ src/__tests__/rocketjumpSelectors.test.js
@@ -41,22 +41,17 @@ describe('Rocketjump selectors', () => {
       state,
       selectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                name: s.name.toUpperCase(),
-                fresh: true,
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              name: s.name.toUpperCase(),
+              fresh: true,
+            }))
           ),
         getOldest: ({ getData }) =>
-          createSelector(
-            getData,
-            soci => {
-              return orderBy(soci, 'age', 'desc')[0]
-            }
-          ),
+          createSelector(getData, soci => {
+            return orderBy(soci, 'age', 'desc')[0]
+          }),
       },
     })()
 
@@ -85,22 +80,17 @@ describe('Rocketjump selectors', () => {
       state,
       proxySelectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                name: s.name.toUpperCase(),
-                fresh: true,
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              name: s.name.toUpperCase(),
+              fresh: true,
+            }))
           ),
         getOldest: ({ getData }) =>
-          createSelector(
-            getData,
-            soci => {
-              return orderBy(soci, 'age', 'desc')[0]
-            }
-          ),
+          createSelector(getData, soci => {
+            return orderBy(soci, 'age', 'desc')[0]
+          }),
       },
     })()
 
@@ -127,13 +117,11 @@ describe('Rocketjump selectors', () => {
     const rjIsAlive = rj({
       selectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                isAlive: s.age < 27,
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              isAlive: s.age < 27,
+            }))
           ),
       },
     })
@@ -142,15 +130,13 @@ describe('Rocketjump selectors', () => {
     const rjRangerName = rj({
       selectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                rangerName: [s.name.slice(0, -2), s.name.slice(-2)]
-                  .map(capitalize)
-                  .join(' '),
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              rangerName: [s.name.slice(0, -2), s.name.slice(-2)]
+                .map(capitalize)
+                .join(' '),
+            }))
           ),
       },
     })
@@ -159,27 +145,23 @@ describe('Rocketjump selectors', () => {
       state,
       selectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                hello: `My name is ${s.rangerName} an i am ${s.age}`,
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              hello: `My name is ${s.rangerName} an i am ${s.age}`,
+            }))
           ),
       },
     })({
       selectors: {
         getData: ({ getData }) =>
-          createSelector(
-            getData,
-            soci =>
-              soci.map(s => ({
-                ...s,
-                hello: s.isAlive
-                  ? `${s.hello} and i am alive`
-                  : `${s.hello} and i am a ghost`,
-              }))
+          createSelector(getData, soci =>
+            soci.map(s => ({
+              ...s,
+              hello: s.isAlive
+                ? `${s.hello} and i am alive`
+                : `${s.hello} and i am a ghost`,
+            }))
           ),
       },
     })
