+++ src/plugins/promise/__tests__/promisePlugin.test.js
@@ -1,7 +1,7 @@
 import configureStore from 'redux-mock-store'
 import createSagaMiddleware from 'redux-saga'
-import { rj } from '../rocketjump'
-import rjWithPromise from '../plugins/promise'
+import { rj } from '../../../rocketjump'
+import rjWithPromise from '../index'
 import { middleware as thunkMiddleware } from 'redux-saga-thunk'
 
 const mockStoreWithSaga = (saga, ...mockStoreArgs) => {
@@ -13,26 +13,25 @@ const mockStoreWithSaga = (saga, ...mockStoreArgs) => {
   return store
 }
 
-describe('Promise catalog', () => {
+describe('Promise plugin', () => {
   it('Should dispatch a promise!', done => {
     const {
       actions: { load },
       saga,
-    } = rj(rjWithPromise,
-      {
-        type: 'OCIO',
-        state: 'X',
-        effect: ({ maik }) =>
-          new Promise((resolve, reject) => {
-            if (maik === 23) {
-              resolve({
-                message: 'Time 2 Rave on!',
-              })
-            } else {
-              reject('Invalid Maik try again')
-            }
-          }),
-      })()
+    } = rj(rjWithPromise, {
+      type: 'OCIO',
+      state: 'X',
+      effect: ({ maik }) =>
+        new Promise((resolve, reject) => {
+          if (maik === 23) {
+            resolve({
+              message: 'Time 2 Rave on!',
+            })
+          } else {
+            reject('Invalid Maik try again')
+          }
+        }),
+    })()
 
     const store = mockStoreWithSaga(saga, {})
     store.dispatch(load({ maik: 777 })).catch(error => {
